+++ downstreamadapter/eventcollector/dispatcher_stat_test.go
@@ -57,7 +57,7 @@ func (m *mockDispatcher) GetStartTs() uint64 {
 	return m.startTs
 }
 
-func (m *mockDispatcher) GetType() int {
+func (m *mockDispatcher) GetType() int64 {
 	return dispatcher.TypeDispatcherEvent
 }
